wip providers resolver working
