+++ apps/cms-v2/src/modules/bulk-sync/bulk-sync-processor.ts
@@ -1,8 +1,4 @@
 import { BulkImportProductFragment } from "../../../generated/graphql";
-import { ProvidersConfig } from "../configuration";
-import { ContentfulBulkSyncProcessor } from "../providers/contentful/contentful-bulk-sync-processor";
-import { DatocmsBulkSyncProcessor } from "../providers/datocms/datocms-bulk-sync-processor";
-import { StrapiBulkSyncProcessor } from "../providers/strapi/strapi-bulk-sync-processor";
 
 export type BulkSyncProcessorHooks = {
   onUploadStart?: (context: { variantId: string }) => void;
@@ -16,19 +12,3 @@ export interface BulkSyncProcessor {
     hooks: BulkSyncProcessorHooks
   ): Promise<void>;
 }
-
-// todo extract to shared
-export const BulkSyncProcessorFactory = {
-  create(config: ProvidersConfig.AnyFullShape): BulkSyncProcessor {
-    switch (config.type) {
-      case "contentful":
-        return new ContentfulBulkSyncProcessor(config);
-      case "datocms":
-        return new DatocmsBulkSyncProcessor(config);
-      case "strapi":
-        return new StrapiBulkSyncProcessor(config);
-      default:
-        throw new Error(`Unknown provider`);
-    }
-  },
-};
