--- apps/cms-v2/src/modules/providers/contentful/index.ts
@@ -1,3 +1,0 @@
-export * from "./contentful";
-export * from "./contentful.router";
-export * from "./contentful-webhooks-processor";
