+++ apps/cms-v2/src/modules/providers/datocms/datocms.tsx
@@ -1,6 +1,6 @@
-import { CMSProvider } from "../../shared/cms-provider";
-import logo from "./datocms.svg";
 import { Box, Text } from "@saleor/macaw-ui/next";
+import { CMSProviderMeta } from "../providers-resolver";
+import logo from "./datocms.svg";
 
 export const Datocms = {
   formSideInfo: (
@@ -17,4 +17,4 @@ export const Datocms = {
   displayName: "DatoCMS",
   description:
     "It's the headless CMS for the modern web. More than 25,000 businesses use DatoCMS to create online content at scale from a central hub and distribute it via API.",
-} satisfies CMSProvider;
+} satisfies CMSProviderMeta;
