--- apps/cms-v2/src/modules/providers/datocms/index.ts
@@ -1,5 +1,0 @@
-export * from "./datocms-client-browser";
-export * from "./datocms-config-form";
-export * from "./datocms-webhooks-processor";
-export * from "./datocms.router";
-export * from "./datocms";
